// src/api/generated/models/zone.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export interface Zone {
  id?: string;
  name: string;
  city: string;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  active?: boolean | null;
  createdAt: string;
  updatedAt: string;
}